+++ backend/config/multer/index.js
@@ -1,0 +1,2 @@
+export { default as uploadDisk } from './multerDiskStorage.js';
+export { default as uploadS3 } from './multerS3Storage.js';
